requirements.txt
Django
python-dotenv
authlib
